put examples in separate package
